Be patient and persistent to work around the bugs when first plugged in.  Check for a 0/0 ID pair and print a reassuring message.  Reworked stl_kick_mode() to reopen more times, and check that the status  returns a valid value before returning.  This takes about 7 seconds of  attempts before we get a valid response.
Implement a new commands.
'erase' as an alias for 'erase=all'
'program=' combines erase with flash write to avoid user errors.

In stl_flash_erase_page(), return 1 if the erase does not work.

When writing flash show a warning message if the binary is larger
than a expected flash size.

If needed, modify then restore the GPIO configuration of the LED pins
when running the "blink" command.

Show CAN registers with "CAN1" and "CAN2" commands.

Show more info with the 'timerN' (e.g. "timer1") commands.